requirements.txt
flask
gunicorn
spacy
jsonpickle
pyenchant